Refactor imagery models and cleanup imports
