+++ rgd/geodata/models/geometry/__init__.py
@@ -1,2 +1,2 @@
-from .base import GeometryArchive, GeometryEntry  # noqa
+from .base import GeometryArchive, GeometryEntry, validate_archive  # noqa
 from .transform import transform_geometry  # noqa
