+++ rgd/geodata/models/imagery/subsample.py
@@ -11,7 +11,7 @@ from rasterio.mask import mask
 from rgd.utility import get_or_create_no_commit
 
 from ..common import ChecksumFile
-from .base import ConvertedImageFile, SubsampledImage
+from .processed import ConvertedImageFile, SubsampledImage
 
 logger = get_task_logger(__name__)
