+++ rgd/geodata/tasks.py
@@ -34,7 +34,7 @@ def _run_with_failure_reason(model, func, *args, **kwargs):
 
 @shared_task(time_limit=86400)
 def task_read_image_file(file_id):
-    from .models.imagery.base import ImageFile
+    from .models.imagery import ImageFile
     from .models.imagery.etl import read_image_file
 
     image_file = ImageFile.objects.get(id=file_id)
@@ -51,7 +51,7 @@ def task_read_geometry_archive(archive_id):
 
 @shared_task(time_limit=86400)
 def task_populate_raster_entry(raster_id):
-    from .models.imagery.base import RasterEntry
+    from .models.imagery import RasterEntry
     from .models.imagery.etl import populate_raster_entry
 
     raster_entry = RasterEntry.objects.get(id=raster_id)
@@ -60,7 +60,7 @@ def task_populate_raster_entry(raster_id):
 
 @shared_task(time_limit=86400)
 def task_populate_raster_footprint(raster_id):
-    from .models.imagery.base import RasterEntry
+    from .models.imagery import RasterEntry
     from .models.imagery.etl import populate_raster_footprint
 
     raster_entry = RasterEntry.objects.get(id=raster_id)
@@ -69,7 +69,7 @@ def task_populate_raster_footprint(raster_id):
 
 @shared_task(time_limit=86400)
 def task_populate_raster_outline(raster_id):
-    from .models.imagery.base import RasterEntry
+    from .models.imagery import RasterEntry
     from .models.imagery.etl import populate_raster_outline
 
     raster_entry = RasterEntry.objects.get(id=raster_id)
@@ -78,7 +78,7 @@ def task_populate_raster_outline(raster_id):
 
 @shared_task(time_limit=86400)
 def task_load_kwcoco_dataset(kwcoco_dataset_id):
-    from .models.imagery.base import KWCOCOArchive
+    from .models.imagery import KWCOCOArchive
     from .models.imagery.etl import load_kwcoco_dataset
 
     ds_entry = KWCOCOArchive.objects.get(id=kwcoco_dataset_id)
@@ -87,7 +87,7 @@ def task_load_kwcoco_dataset(kwcoco_dataset_id):
 
 @shared_task(time_limit=86400)
 def task_read_fmv_file(file_id):
-    from .models.fmv.base import FMVFile
+    from .models.fmv import FMVFile
     from .models.fmv.etl import read_fmv_file
 
     fmv_file = FMVFile.objects.get(id=file_id)
@@ -96,7 +96,7 @@ def task_read_fmv_file(file_id):
 
 @shared_task(time_limit=86400)
 def task_convert_to_cog(conv_id):
-    from .models.imagery.base import ConvertedImageFile
+    from .models.imagery import ConvertedImageFile
     from .models.imagery.subsample import convert_to_cog
 
     cog = ConvertedImageFile.objects.get(id=conv_id)
@@ -105,7 +105,7 @@ def task_convert_to_cog(conv_id):
 
 @shared_task(time_limit=86400)
 def task_populate_subsampled_image(subsampled_id):
-    from .models.imagery.base import SubsampledImage
+    from .models.imagery import SubsampledImage
     from .models.imagery.subsample import populate_subsampled_image
 
     cog = SubsampledImage.objects.get(id=subsampled_id)
