+++ rgd/geodata/tests/test_fmv.py
@@ -3,7 +3,7 @@ import os
 import pytest
 
 from rgd.geodata.datastore import datastore
-from rgd.geodata.models.fmv.base import FMVEntry
+from rgd.geodata.models.fmv import FMVEntry
 from rgd.geodata.models.mixins import Status
 
 from . import factories
