+++ rgd/geodata/tests/test_geometry.py
@@ -22,4 +22,4 @@ def test_geometry_etl(sample_file):
     # Testing that we can repopulate a geometry archive again
     read_geometry_archive(geom_archive.id)
     # test the field file validator
-    models.geometry.base.validate_archive(geom_archive.file.file)
+    models.geometry.validate_archive(geom_archive.file.file)
